refactor: create service class for seller quotes
